Made methods for drawText, drawBalloon, and drawCheckered (also fixed some bugs)

+++ app/src/main/java/cs301/birthdaycake/CakeModel.java
@@ -57,12 +57,11 @@ public class CakeModel
 
     public int getCoordY() { return CoordY; }
 
-
     public void setX(int x) { this.x = x; }
 
-    public float getX() { return x; }
+    public int getX() { return x; }
 
     public void setY(int y) { this.y = y; }
 
-    public float getY() { return y; }
+    public int getY() { return y; }
 }
